Level generation with multiple room assets
Experimented with using public game object list to store multiple room assets to enable level generation with multiple different rooms. In the future, it is considered to use multiple such list to store rooms of different direction and roads would no longer be needed. The generation of special room is still under consideration.