Shadow Doodle
In the TestScene, the shadowDoodle is ready to be tested. This enemy will disappear after 10 seconds after it has been born. However, once the player enters its detection zone (a radius of 5), it will disappear and the player would be blinded with a pee meter increase of 20. Another thing to be implemented is that once spawn, the shadowDoodle should find the PeeMeter in the scene and initialize that variable in its script.