Fairy Prototype
Implemented Fairy and the tree. Now, the fairy will try to mix its position in the middle of the screen and keep a constant distance away from the player (to make this distance fixed or varied is yet to be discussed), and it will scare the player once it collides with any other objects (including ground and wall, maybe we can make them safe in the future). If it collides with the tree, they both disappear and player's peemeter would be reduced. Spawning mechanism to be implemented