Few changes to the enemy spawn and some enemy feature
Ignoring changing spawning mechanism for now
Done(maybe). I would need to revert the spawn mechanism for a reasonable playtest
Somehow the nontrigger ShadowDoodle collider disappeared, I have added it back. I have playtested it and shadowdoodle generated normally after I made its spawn condition less strict
Done. Basically modified the code of the ghost and make the look threshold stricter so that our player need to almost put it to the center of the visual cone to trigger it